video-seal/requirements.txt
fastapi
uvicorn
httpx
cloudinary
ffmpeg-python
torch
opencv-python
numpy